requirements.txt
streamlit==1.25.0
fastapi==0.100.0
uvicorn==0.23.1
langchain==0.0.236
openai==0.27.8
fpdf==1.7.2
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
google-search-results==2.4.2
pydantic<2.0.0
